+++ packages/browser/src/image/convert.ts
@@ -4,6 +4,7 @@ import {
   elementToCanvas2d,
   getBlobAsync,
 } from "../canvas/convert";
+import { bytesToBase64 } from "@a2-ts-utils/common/byte";
 
 export function getElementImageBlobAsync(
   element: HTMLVideoElement | HTMLImageElement | HTMLCanvasElement,
@@ -21,4 +22,12 @@ export async function getElementImageBytesAsync(
 ): Promise<Uint8Array> {
   var blob = await getBlobAsync(elementToCanvas2d(element, newSize), type);
   return blobToBytesAsync(blob);
+}
+
+export function bytesToBase64Src(
+  bytes: Uint8Array,
+  type: ImageMimeType = ImageMimeType.JPEG,
+): string {
+  const base64String = bytesToBase64(bytes);
+  return `data:${type};base64,${base64String}`;
 }