+++ packages/common/src/index.ts
@@ -1,5 +1,6 @@
 export * as task from "./task";
 export * as image from "./image";
 export * as type from "./type";
+export * as byte from "./byte";
 
 export const Name = "@a2-ts-utils/common";