+++ packages/browser/history.md
@@ -2,6 +2,7 @@
 
 ## 26-04-29(wed)
 
+- bytesToBase64Src を追加
+- VideoやImageのリサイズ処理を追加
 - 画像の変換処理を分割及びカスタマイズ
   - element > canvas > blob > buffer > bytes
-- VideoやImageのリサイズ処理を追加